requirements.txt
numpy==2.1.0
pillow==10.4.0
scipy==1.14.0
setuptools==72.2.0
